# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Clinic Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get today's dashboard statistics
      parameters:
        - name: branch
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /dashboard/funnel:
    get:
      operationId: getDailyFunnel
      tags: [dashboard]
      summary: Get daily appointment status funnel
      parameters:
        - name: date
          in: query
          schema:
            type: ["string", "null"]
        - name: branch
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Funnel data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyFunnel"

  # ─────────────────────────────────────────────
  # PATIENTS
  # ─────────────────────────────────────────────
...
  /financial/summary:
    get:
      operationId: getFinancialSummary
      tags: [financial]
      summary: Get financial KPI summary
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
        - name: branch
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Financial summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialSummary"
  /financial/receivables:
    get:
      operationId: listReceivables
      tags: [financial]
      summary: List outstanding receivables
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Receivables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Receivable"
...
  /financial/vaults/{id}/transactions:
    get:
      operationId: listVaultTransactions
      tags: [financial]
      summary: Get vault transaction history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VaultTransaction"
    post:
      operationId: createVaultTransaction
      tags: [financial]
      summary: Create a vault transaction (deposit/withdraw/transfer)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultTransactionInput"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultTransaction"
  /financial/expenses:
    get:
      operationId: listExpenses
      tags: [financial]
      summary: List expenses
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [financial]
      summary: Record a new expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /financial/expense-categories:
    get:
      operationId: listExpenseCategories
      tags: [financial]
      summary: List expense categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseCategory"
    post:
      operationId: createExpenseCategory
      tags: [financial]
      summary: Create expense category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseCategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseCategory"

  # ─────────────────────────────────────────────
  # SERVICES
  # ─────────────────────────────────────────────
...
  /inventory/items:
    get:
      operationId: listInventoryItems
      tags: [inventory]
      summary: List inventory items
      parameters:
        - name: branch
          in: query
          schema:
            type: ["string", "null"]
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: lowStock
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Add a new inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/items/{id}/transaction:
    post:
      operationId: inventoryTransaction
      tags: [inventory]
      summary: Add or withdraw inventory item quantity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryTransactionInput"
      responses:
        "201":
          description: Transaction recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryTransaction"
...
components:
  schemas:
...
    DashboardStats:
      type: object
      properties:
        todayAppointments:
          type: integer
        newPatientsToday:
          type: integer
        totalPatients:
          type: integer
        todayRevenue:
          type: number
        totalVisits:
          type: integer
        totalAppointments:
          type: integer
      required: [todayAppointments, newPatientsToday, totalPatients, todayRevenue, totalVisits, totalAppointments]
    DailyFunnel:
      type: object
      properties:
        waitingArrival:
          type: integer
        inReception:
          type: integer
        inExamination:
          type: integer
        completed:
          type: integer
        postponed:
          type: integer
        noShow:
          type: integer
        sessionDone:
          type: integer
      required: [waitingArrival, inReception, inExamination, completed, postponed, noShow, sessionDone]

    # ── Patient ──
...
    FinancialSummary:
      type: object
      properties:
        totalPayments:
          type: number
        totalReceivables:
          type: number
        totalDiscounts:
          type: number
        totalServicePrice:
          type: number
        totalPatients:
          type: integer
        totalServices:
          type: integer
        netProfit:
          type: number
        totalExpenses:
          type: number
        cashPayments:
          type: number
        otherPayments:
          type: number
      required: [totalPayments, totalReceivables, totalDiscounts, totalServicePrice, totalPatients, totalServices, netProfit, totalExpenses, cashPayments, otherPayments]
    Receivable:
      type: object
      properties:
        patientId:
          type: integer
        patientName:
          type: string
        patientPhone:
          type: ["string", "null"]
        appointmentId:
          type: integer
        appointmentDate:
          type: string
        amount:
          type: number
      required: [patientId, patientName, appointmentId, appointmentDate, amount]
...
    VaultTransaction:
      type: object
      properties:
        id:
          type: integer
        vaultId:
          type: integer
        type:
          type: string
          enum: [deposit, withdrawal, transfer_in, transfer_out]
        amount:
          type: number
        note:
          type: ["string", "null"]
        performedBy:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, vaultId, type, amount, createdAt]
    VaultTransactionInput:
      type: object
      properties:
        type:
          type: string
          enum: [deposit, withdrawal, transfer_in, transfer_out]
        amount:
          type: number
        targetVaultId:
          type: ["integer", "null"]
        note:
          type: ["string", "null"]
      required: [type, amount]
    Expense:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: string
        amount:
          type: number
        vaultId:
          type: integer
        vaultName:
          type: string
        performedBy:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        receiptUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, categoryId, categoryName, amount, vaultId, vaultName, createdAt]
    ExpenseInput:
      type: object
      properties:
        categoryId:
          type: integer
        amount:
          type: number
        vaultId:
          type: integer
        performedBy:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        receiptUrl:
          type: ["string", "null"]
      required: [categoryId, amount, vaultId]
    ExpenseCategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
      required: [id, name]
    ExpenseCategoryInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
      required: [name]

    # ── Services ──
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        barcode:
          type: ["string", "null"]
        branch:
          type: string
        name:
          type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: string
        lowStockThreshold:
          type: ["number", "null"]
        expiryDate:
          type: ["string", "null"]
        supplierName:
          type: ["string", "null"]
        supplierContact:
          type: ["string", "null"]
        notifyLowStock:
          type: boolean
        notifyExpiry:
          type: boolean
      required: [id, branch, name, quantity, unit, notifyLowStock, notifyExpiry]
    InventoryItemInput:
      type: object
      properties:
        barcode:
          type: ["string", "null"]
        branch:
          type: string
        name:
          type: string
        categoryId:
          type: ["integer", "null"]
        quantity:
          type: number
        unit:
          type: string
        lowStockThreshold:
          type: ["number", "null"]
        expiryDate:
          type: ["string", "null"]
        supplierName:
          type: ["string", "null"]
        supplierContact:
          type: ["string", "null"]
        supplierAddress:
          type: ["string", "null"]
        notifyLowStock:
          type: boolean
        notifyExpiry:
          type: boolean
      required: [branch, name, quantity, unit, notifyLowStock, notifyExpiry]
    InventoryTransactionInput:
      type: object
      properties:
        type:
          type: string
          enum: [add, withdraw]
        quantity:
          type: number
        note:
          type: ["string", "null"]
        cost:
          type: ["number", "null"]
      required: [type, quantity]
    InventoryTransaction:
      type: object
      properties:
        id:
          type: integer
        itemId:
          type: integer
        type:
          type: string
        quantity:
          type: number
        note:
          type: ["string", "null"]
        cost:
          type: ["number", "null"]
        createdAt:
          type: string
      required: [id, itemId, type, quantity, createdAt]
...