# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Clinic Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  categoryId: number;
  categoryName: string;
  amount: number;
  vaultId: number;
  vaultName: string;
  /** @nullable */
  performedBy?: string | null;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  receiptUrl?: string | null;
  createdAt: string;
}