# lib/api-zod/src/generated/types/vaultTransaction.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Clinic Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { VaultTransactionType } from './vaultTransactionType';

export interface VaultTransaction {
  id: number;
  vaultId: number;
  type: VaultTransactionType;
  amount: number;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  performedBy?: string | null;
  createdAt: string;
}